+++ include/constant.hpp
@@ -2,8 +2,8 @@
 #include "fpr.hpp"
 
 class Z : public FPR {
-  void exec(std::vector<int> n) {
-    result.resize(0); // no me fio
+  void exec(std::vector<long int  > n) {
+    result.resize(0); 
     result.push_back(0);
   }
 };
